+++ stdlib/LinearAlgebra/src/LinearAlgebra.jl
@@ -102,7 +102,7 @@ export
     istriu,
     kron,
     ldiv!,
-    ldltfact!,
+    ldlt!,
     ldlt,
     linreg,
     logabsdet,