+++ NEWS.md
@@ -703,8 +703,9 @@ Deprecated or removed
   * `lufact`, `schurfact`, `lqfact`, `qrfact`, and `ldltfact` have respectively
     been deprecated to `lu`, `schur`, `lq`, `qr`, and `ldlt` ([#27159]).
 
-  * `lufact!`, `schurfact!`, `lqfact!`, and `qrfact!` have
-    respectively been deprecated to `lu!`, `schur!`, `lq!`, and `qr!` ([#27159]).
+  * `lufact!`, `schurfact!`, `lqfact!`, `qrfact!`,
+    and `ldltfact!` have respectively been deprecated to `lu!`,
+    `schur!`, `lq!`, `qr!`, and `ldlt!` ([#27159]).
 
   * Indexing into multidimensional arrays with more than one index but fewer indices than there are
     dimensions is no longer permitted when those trailing dimensions have lengths greater than 1.